feat: display markers and on click nav

+++ app/src/main/java/com/tapandgo/mybikes/model/Station.kt
@@ -17,7 +17,7 @@ data class Station (
 ) {
 
     class GeoPosition(
-        val lat: Float,
-        val lng: Float,
+        val lat: Double,
+        val lng: Double,
     )
 }
